requirements.txt
mysql_connector_repackaged==0.3.1
numpy==1.24.2
pandas==1.5.3
streamlit==1.20.0
streamlit_extras==0.2.7
